+++ R/predictions.R
@@ -78,7 +78,7 @@ select_prediction_method <- function(model_class,
     prediction_data <- get_predictions_zelig(model, data_grid, ci.lvl, linv, ...)
   } else if (model_class == "polr") {
     prediction_data <- get_predictions_polr(model, data_grid, ci.lvl, linv, value_adjustment, terms, model_class, vcov.fun, vcov.type, vcov.args, condition, interval, ...)
-  } else if (model_class %in% c("betareg", "truncreg", "ivreg", "vgam")) {
+  } else if (model_class %in% c("betareg", "truncreg", "ivreg", "vgam", "fixest")) {
     prediction_data <- get_predictions_generic2(model, data_grid, ci.lvl, linv, type, model_class, value_adjustment, terms, vcov.fun, vcov.type, vcov.args, condition, interval, ...)
   } else if (model_class %in% c("zeroinfl", "hurdle", "zerotrunc")) {
     prediction_data <- get_predictions_zeroinfl(model, data_grid, ci.lvl, linv, type, model_class, value_adjustment, terms, vcov.fun, vcov.type, vcov.args, condition, ...)